Move rebased code from DataTransform to DataFetch
