+++ SharePointFramework/ProjectWebParts/src/components/ProjectInformation/useProjectInformationDataFetch.ts
@@ -1,5 +1,6 @@
 import SPDataAdapter from 'data'
 import { ProjectAdminPermission } from 'data/SPDataAdapter/ProjectAdminPermission'
+import strings from 'ProjectWebPartsStrings'
 import { useEffect } from 'react'
 import { isEmpty } from 'underscore'
 import { ProjectPropertyModel } from './ProjectProperties/ProjectProperty'
@@ -31,6 +32,41 @@ const transformProperties = (
   return properties
 }
 
+const projectDataSynced = async (
+  props: IProjectInformationProps
+) => {
+  try {
+    let isSynced = false
+
+    const projectDataList = props.hubSite.web.lists
+      .getByTitle(strings.IdeaProjectDataTitle)
+
+    const [projectDataItem] = await projectDataList
+      .items
+      .filter(`GtSiteUrl eq '${props.webPartContext.pageContext.web.absoluteUrl}'`)
+      .select('Id')
+      .get()
+
+    const ideaProcessingList = props.hubSite.web.lists.getByTitle(strings.IdeaProcessingTitle)
+
+    const [ideaProcessingItem] = await ideaProcessingList
+      .items
+      .filter(`GtIdeaProjectDataId eq '${projectDataItem.Id}'`)
+      .select('Id, GtIdeaDecision')
+      .get()
+
+    if (ideaProcessingItem.GtIdeaDecision === 'Godkjent og synkronisert') {
+      isSynced = true
+    }
+
+    console.log('isDataSynced', isSynced)
+
+    return isSynced
+  } catch (error) {
+    return false
+  }
+}
+
 const fetchData = async (
   props: IProjectInformationProps
 ): Promise<Partial<IProjectInformationState>> => {
@@ -49,12 +85,14 @@ const fetchData = async (
     )
     const properties = transformProperties(data, props)
     const allProperties = transformProperties(data, props, false)
+    const isProjectDataSynced = props.useIdeaProcessing && await projectDataSynced(props)
     return {
       data,
       isParentProject: data.fieldValues?.GtIsParentProject || data.fieldValues?.GtIsProgram,
       properties,
       allProperties,
-      userHasEditPermission
+      userHasEditPermission,
+      isProjectDataSynced
     }
   } catch (error) {
     throw error
